// gclampv3/student/src/app/services/enum.ts
export enum fldNames {
    cifcnt_fld = "cifcnt_fld",
    piccnt_fld = "piccnt_fld",
    ciccnt_fld = "ciccnt_fld",
    taskcnt_fld = "taskcnt_fld",
    class_on_time_fld = "class_on_time_fld",
    items_acquired_fld = "items_acquired_fld",
    points_fld = "points_fld"
}

export enum Crud {
    ADD,UPDATE,DELETE,VIEW,REMOVE,CREATE,READ
}

export enum ControlForAddingPoints {
    BRONZE = 9,
    SILVER = 49,
    GOLD = 99,
    PLATINUM = 149
}

export enum ListOfAchievementTargets {
    BRONZE = 10,
    SILVER = 50,
    GOLD = 100,
    PLATINUM = 150
}